+++ git-protocol/tests/fetch/mod.rs
@@ -15,7 +15,7 @@ type Cursor = std::io::Cursor<Vec<u8>>;
 type Cursor = futures_lite::io::Cursor<Vec<u8>>;
 
 #[allow(clippy::result_large_err)]
-fn helper_unused(_action: git_credentials::helper::Action) -> git_credentials::protocol::Result {
+fn helper_unused(_action: gix_credentials::helper::Action) -> gix_credentials::protocol::Result {
     panic!("Call to credentials helper is unexpected")
 }