+++ git-transport/src/client/blocking_io/http/curl/mod.rs
@@ -31,7 +31,7 @@ pub enum Error {
     #[error("Could not finish reading all data to post to the remote")]
     ReadPostBody(#[from] std::io::Error),
     #[error(transparent)]
-    Authenticate(#[from] git_credentials::protocol::Error),
+    Authenticate(#[from] gix_credentials::protocol::Error),
 }
 
 impl crate::IsSpuriousError for Error {